use floats not ints in convolve, move convolve test to appropriate test file
